py-processor/requirements-simple.txt
# Simplified requirements for Render deployment
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
boto3==1.34.0
minio==7.2.0
python-multipart==0.0.6
numpy==1.24.3
requests==2.31.0
google-generativeai==0.3.2
redis==5.0.1
setuptools>=65.0.0
wheel>=0.38.0
